requirements.txt
Django==1.8.13
django-compressor==1.6
django-libsass==0.7
